Modularize directory structure.
