+++ client.src/core/views/footer-view/footer-view.jst
