+++ client.src/core/views/footer-view/index.js
