+++ client.src/core/views/menu-views/auth-menu-view/index.js
