+++ client.src/core/views/menu-views/auth-menu-view/menu-view.jst
