+++ client.src/lib/routing/router.js
@@ -1,7 +1,6 @@
 /* jshint maxstatements: 20 */
 
 import * as _ from 'underscore';
-import * as mn from 'marionette';
 import * as Radio from 'radio';
 import * as BaseRouter from 'backbone.base-router';
 import Route from './route';
