+++ client.src/modules/about/views/about-view/index.js
