+++ client.src/modules/entities/gistbook-page/index.js
