+++ client.src/modules/entities/github-collection/index.js
