+++ client.src/modules/entities/github-user/index.js
