+++ client.src/modules/logout/logout-route.js
@@ -2,9 +2,8 @@
 // LogoutRoute
 //
 
-import * as mn from 'marionette';
 import * as Radio from 'radio';
-import Route from '../lib/routing/route';
+import Route from '../../lib/routing/route';
 
 var authChannel = Radio.channel('auth');
