+++ client.src/modules/profile/gistbook-route.js
@@ -2,12 +2,11 @@
 // GistbookRoute
 //
 
-import * as mn from 'marionette';
 import * as Radio from 'radio';
-import Route from '../../../lib/routing/route';
-import GistView from '../../../features/views/gist-view';
-import Gist from '../../../features/entities/gist';
-import ServerErrorView from '../../../features/views/server-error-view';
+import Route from '../../lib/routing/route';
+import GistView from '../../shared/views/gist-view';
+import Gist from '../entities/gist';
+import ServerErrorView from '../../shared/views/server-error-view';
 
 export default Route.extend({
   fetch: function(urlData) {
