+++ client.src/modules/profile/profile-route.js
@@ -2,13 +2,12 @@
 // ProfileRoute
 //
 
-import * as mn from 'marionette';
 import * as Radio from 'radio';
 import Route from '../../lib/routing/route';
-import ProfileView from '../../features/views/profile-view';
-import Gists from '../../features/views/profile-view/entities/gists';
-import GithubUser from '../../features/entities/github-user';
-import ServerErrorView from '../../features/views/server-error-view';
+import ProfileView from './views/profile-view';
+import Gists from './entities/gists';
+import GithubUser from '../entities/github-user';
+import ServerErrorView from '../../shared/views/server-error-view';
 
 export default Route.extend({
   fetch: function(urlData) {
