+++ client.src/modules/settings/settings-route.js
@@ -2,10 +2,9 @@
 // SettingsRoute
 //
 
-import * as mn from 'marionette';
 import * as Radio from 'radio';
-import Route from '../lib/routing/route';
-import SettingsView from '../features/views/settings-view';
+import Route from '../../lib/routing/route';
+import SettingsView from './views/settings-view';
 
 var user = Radio.request('user', 'user');
