+++ client.src/modules/settings/views/revoke-modal/index.js
