+++ client.src/modules/settings/views/revoke-modal/revoke-modal.styl
