+++ client.src/modules/settings/views/settings-view/index.js
@@ -5,8 +5,8 @@
 import * as _ from 'underscore';
 import * as mn from 'marionette';
 import * as Radio from 'radio';
-import RevokeModalView from './revoke-modal';
-import scopeMap from '../../../util/scope-map';
+import RevokeModalView from '../revoke-modal';
+import scopeMap from '../../../../util/scope-map';
 
 export default mn.ItemView.extend({
   template: 'settingsView',
