+++ client.src/modules/settings/views/settings-view/settings-view.jst
