+++ client.src/modules/terms/views/terms-view/index.js
