+++ client.src/modules/terms/views/terms-view/terms-view.jst
