+++ client.src/modules/terms/views/terms-view/terms.styl
