+++ client.src/shared/views/gist-view/gist-view.styl
