+++ client.src/shared/views/server-error-view/index.js
