+++ client.src/shared/views/server-error-view/server-error-view.jst
